feat(OCMUI-2905): Connect Windows LI UI components to related api queries
Side note: I believe these are the necessary changes needed for the Machine Pool to be created with Windows Licensing. We will not be able to know for sure until the api will be exposed on stage, and we can test and verify that.
These changes contain some comments to help others understand the changes required, they should not be included in the final PR

Link to issue: https://issues.redhat.com/browse/OCMUI-2905